stop depending on attachments
While it's cool to rebuild a note every time its attachment is changed, it is highly costly operation. On average, it
eats around 20 seconds on the set of 245 notes. And since I never change attachment content, this penalty is doesn't
worth it. I love my `make notes` fast (e.g. 1 second when there are no changes).